+++ src/cmd/cmd_conc.mli
@@ -13,6 +13,7 @@ val cmd :
   -> bool
   -> bool
   -> bool
+  -> bool
   -> Cmd_sym.fail_mode
   -> Fpath.t
   -> Smtml.Solver_dispatcher.solver_type