+++ test/c/eacsl/ghost/else.t
@@ -1,4 +1,7 @@
-  $ owi c --e-acsl ./else.c --no-value 2>1 | grep -v "Assert failure"
+  $ owi c --e-acsl ./else.c --no-value --no-assert-failure-expression-printing
+  Assert failure
   Model:
     (model
       (symbol_0 i32))
+  Reached problem!
+  [13]